requirements.txt
mysql-connector-python==8.0.25
PySimpleGUI==4.40.0